# db-sync-service/docs/swagger.yaml
basePath: /api
definitions:
  healthz.Status:
    properties:
      api-name:
        type: string
      status:
        type: string
    type: object
  models.Contact:
    properties:
      city:
        type: string
      country:
        type: string
      email:
        type: string
      firstname:
        type: string
      home:
        type: string
      id:
        type: integer
      lastname:
        type: string
      mobile:
        type: string
      postalCode:
        type: string
      province:
        type: string
      street:
        type: string
      work:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  title: User API documentation
  version: 1.0.0
paths:
  /contacts:
    get:
      consumes:
      - application/json
      description: Get all Contacts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Contact'
              type: array
            type: object
      summary: Get all Contacts
      tags:
      - GetContacts
  /contacts/{id}:
    delete:
      consumes:
      - application/json
      description: delete contact by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a Contact
      tags:
      - DeleteContact
    get:
      consumes:
      - application/json
      description: get contact by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Find a Contact
      tags:
      - FindContact
    patch:
      consumes:
      - application/json
      description: Update contact by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contact
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a Contact
      tags:
      - UpdateContact
  /contacts/create:
    post:
      consumes:
      - application/json
      description: get contact by ID
      parameters:
      - description: Contact
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a Contact
      tags:
      - CreateContact
  /healthz:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthz.Status'
      summary: Show the status of address nook server.
      tags:
      - ServerStatus
swagger: "2.0"